feat: major improvement-1

+++ InventoryManagement/Models/Item.cs
@@ -4,7 +4,7 @@ namespace InventoryManagement.Models.Item {
     class Item {
         public string itemCode { get; set; }
         public string itemName { get; set; }
-        public int itemQuantity { get; set; }
+        public double itemQuantity { get; set; }
         public string itemNote { get; set; }
     }
 }
